pyproject.toml
[project]
name = "content-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "genanki>=0.13.1",
    "langchain>=1.2.6",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "loguru>=0.7.3",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
]
